pyproject.toml
[project]
name = "smart-answers"
version = "0.2.0"
description = ""
authors = [
    {name = "Leonn-Leon",email = "skovoroda.vitalik@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "flask-core (>=2.9.0,<3.0.0)",
    "flask-cors (>=5.0.0,<6.0.0)",
    "gigachain (>=0.2.16,<0.3.0)",
    "gigachain-community (>=0.2.16.post3,<0.3.0)",
    "unstructured[all-docs] (>=0.16.21,<0.17.0)",
    "faiss-cpu (>=1.10.0,<2.0.0)",
    "sentence-transformers (>=3.4.1,<4.0.0)",
    "langchain-community (>=0.3.17,<0.4.0)",
    "langchain-gigachat (>=0.3.4,<0.4.0)",
    "scikit-learn (>=1.6.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
